@@ -3,28 +3,24 @@
   "version": "1.0.0",
   "main": "index.js",
   "private": true,
+  "type": "module",
   "author": "Alvin",
   "license": "ISC",
   "scripts": {
-    "lint": "eslint src --ext js --fix",
+    "lint": "eslint src/**/*.js --fix",
     "test": "mocha src/**/*.spec.js --color"
   },
   "dependencies": {
-    "fs-extra": "^11.1.1",
-    "glob": "^10.2.7",
-    "jsdom": "^22.1.0",
+    "fs-extra": "^11.2.0",
+    "glob": "^11.0.0",
+    "jsdom": "^25.0.1",
     "promise": "^8.3.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.23.3",
-    "chai": "^4.3.7",
-    "eslint": "^8.55.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.0.1",
-    "eslint-plugin-promise": "^6.1.1",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
-    "mocha": "^10.2.0"
+    "chai": "^5.1.2",
+    "eslint": "^9.16.0",
+    "globals": "^15.12.0",
+    "mocha": "^10.8.2"
   },
-  "packageManager": "pnpm@8.15.5+sha512.b051a32c7e695833b84926d3b29b8cca57254b589f0649d899c6e9d0edb670b91ec7e2a43459bae73759bb5ce619c3266f116bf931ce22d1ef1759a7e45aa96f"
+  "packageManager": "pnpm@9.14.4+sha512.c8180b3fbe4e4bca02c94234717896b5529740a6cbadf19fa78254270403ea2f27d4e1d46a08a0f56c89b63dc8ebfd3ee53326da720273794e6200fcf0d184ab"
 }
