@@ -1,6 +1,7 @@
-const { expect } = require('chai');
-const { describe, it } = require('mocha');
-const { EventEmitter } = require('events');
+import { describe, it } from 'mocha';
+import { expect } from 'chai';
+
+import { EventEmitter } from 'node:events';
 
 /**
  * 测试异步调用
@@ -21,7 +22,7 @@ describe('test async call', () => {
   /**
    * 回调方式进行异步调用
    */
-  it('should call callback functions', done => {
+  it('should call callback functions', (done) => {
     /**
      * 定义异步调用函数
      */
@@ -39,30 +40,38 @@ describe('test async call', () => {
     let count = 0;
 
     // 异步调用, 回调 success 函数
-    after(true, 50,
-      m => {
+    after(
+      true,
+      50,
+      (m) => {
         expect(m).is.eq('OK');
         count++;
       },
-      e => done(e));
+      (e) => done(e)
+    );
 
     // 异步调用, 回调 error 函数
-    after(false, 50,
-      m => done(m),
-      e => {
+    after(
+      false,
+      50,
+      (m) => done(m),
+      (e) => {
         expect(e).is.eq('Error');
         count++;
       }
     );
 
     // 等待调用结束
-    wait(() => count === 2, () => done());
+    wait(
+      () => count === 2,
+      () => done()
+    );
   });
 
   /**
    * 以事件作为异步函数的回调
    */
-  it('should emit event', done => {
+  it('should emit event', (done) => {
     // 定义事件对象
     const emitter = new EventEmitter();
 
@@ -83,13 +92,13 @@ describe('test async call', () => {
     let count = 0;
 
     // 定义 success 事件监听函数
-    emitter.on('success', m => {
+    emitter.on('success', (m) => {
       expect(m).is.eq('OK');
       count++;
     });
 
     // 定义 error 事件监听函数
-    emitter.on('error', e => {
+    emitter.on('error', (e) => {
       expect(e).is.eq('Error');
       count++;
     });
@@ -101,10 +110,13 @@ describe('test async call', () => {
     after(false, 50);
 
     // 等待调用结束
-    wait(() => count === 2, () => done());
+    wait(
+      () => count === 2,
+      () => done()
+    );
   });
 
-  it('should callback by `Promise`', done => {
+  it('should callback by `Promise`', (done) => {
     // 定义异步函数, 返回 Promise 对象
     const after = (ok, timeout) =>
       // 返回 Promise 对象
@@ -125,22 +137,25 @@ describe('test async call', () => {
 
     // 调用异步函数, 并最终回调 Promise 对象的 then 函数
     after(true, 50)
-      .then(m => {
+      .then((m) => {
         expect(m).is.eq('OK');
         count++;
       })
-      .catch(e => done(e));
+      .catch((e) => done(e));
 
     // 调用异步函数, 并最终回调 Promise 对象的 catch 函数
     after(false, 50)
-      .then(m => done(m))
-      .catch(e => {
+      .then((m) => done(m))
+      .catch((e) => {
         expect(e.message).is.eq('Error');
         count++;
       });
 
     // 等待调用结束
-    wait(() => count === 2, () => done());
+    wait(
+      () => count === 2,
+      () => done()
+    );
   });
 
   /**
@@ -161,9 +176,7 @@ describe('test async call', () => {
             reject(new Error('Error'));
           }
         }, timeout);
-      })
-      ;
-
+      });
     // 等待异步函数调用完毕并返回正确结果
     const m = await after(true, 50);
     expect(m).is.eq('OK');
@@ -179,7 +192,7 @@ describe('test async call', () => {
   /**
    * 合并多个 Promise 对象并行调用
    */
-  it('merge all `Promise` object together', done => {
+  it('merge all `Promise` object together', (done) => {
     // 定义异步函数, 返回 Promise 对象
     const after = (ok, timeout) =>
       // 返回 Promise 对象
@@ -193,25 +206,25 @@ describe('test async call', () => {
             reject(new Error('Error'));
           }
         }, timeout);
-      })
-      ;
-
+      });
     // 合并多个 Promise 对象
     const promise = Promise.all([
       after(true, 50), // 2 个正确调用
       after(true, 50),
-      after(false, 50) // 1 个错误调用
+      after(false, 50), // 1 个错误调用
     ]);
 
     // 查看所有 Promise 对象执行结果
     promise
-      .then(rs => { // 2 个正确结果
+      .then((rs) => {
+        // 2 个正确结果
         expect(rs).to.have.length(2);
-        expect(rs.every(m => m === 'OK')).is.true;
+        expect(rs.every((m) => m === 'OK')).is.true;
       })
-      .catch(es => { // 1 个错误结果
+      .catch((es) => {
+        // 1 个错误结果
         expect(es).to.have.length(1);
-        expect(es.every(e => e === 'Error')).is.true;
+        expect(es.every((e) => e === 'Error')).is.true;
       })
       .finally(() => done());
   });
@@ -234,14 +247,12 @@ describe('test async call', () => {
             reject(new Error('Error'));
           }
         }, timeout);
-      })
-      ;
-
+      });
     // 合并多个 Promise 对象并等待全部的返回结果
     const rs = await Promise.all([
       after(true, 50), // 3 个正确调用
       after(true, 50),
-      after(true, 50)
+      after(true, 50),
     ]);
 
     expect(rs).to.have.length(3);
@@ -251,7 +262,7 @@ describe('test async call', () => {
       await Promise.all([
         after(true, 50), // 包含一个错误调用
         after(true, 50),
-        after(false, 50)
+        after(false, 50),
       ]);
     } catch (e) {
       // 任意返回错误的 Promise 都会导致异常, 造成整个调用全部失败
