@@ -1,7 +1,11 @@
-const { describe, it } = require('mocha');
-const { expect } = require('chai');
-const { JSDOM } = require('jsdom');
-const { startServer, get, post } = require('.');
+
+import { describe, it } from 'mocha';
+import { expect } from 'chai';
+
+import { JSDOM } from 'jsdom';
+
+import * as client from './client.js';
+import startServer from './server.js';
 
 /**
  * 测试 HTTP 服务和客户端
@@ -12,11 +16,11 @@ describe('test `http` module', () => {
    */
   it('should create http server and can be visited by client', async () => {
     // 实例化 HTTP 服务端
-    const server = await startServer(3001);
+    const srv = await startServer(3001);
 
     try {
       // 通过 GET 请求访问服务端 / 地址, 确认返回响应正确
-      let resp = await get('http://localhost:3001');
+      let resp = await client.get('http://localhost:3001');
       expect(resp.code).is.eq(200);
       expect(resp.headers).has.property('content-type', 'text/html; charset=utf-8');
       expect(resp.headers).has.property('content-length', '613');
@@ -27,7 +31,7 @@ describe('test `http` module', () => {
       expect(content).is.eq('Hello World');
 
       // 通过 POST 请求访问服务端 /d/version 地址
-      resp = await post('http://localhost:3001/d/version');
+      resp = await client.post('http://localhost:3001/d/version');
       expect(resp.code).is.eq(200);
       expect(resp.headers).has.property('content-type', 'application/json; charset=utf-8');
       expect(resp.headers).has.property('content-length', '31');
@@ -37,7 +41,7 @@ describe('test `http` module', () => {
       expect(json).has.property('version', '1.0.0');
       expect(json).has.property('build', 101);
     } finally {
-      server.shutdown();
+      srv.shutdown();
     }
   });
 });
