@@ -1,6 +1,1 @@
-const { Person, add } = require('./person');
-
-// 导出类型和函数
-module.exports = {
-  Person, add
-};
+export { Person, add } from './person.js';
